requirements.txt
asgiref==3.8.1
boto3==1.35.4
botocore==1.35.4
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Deprecated==1.2.14
distro==1.9.0
Django==5.1
django-allauth==64.1.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-storages==1.14.4
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
gunicorn==23.0.0
idna==3.7
jmespath==1.0.1
netaddr==1.3.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pip-chill==1.0.3
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
yookassa==3.3.0